# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Music AI App API
  version: 0.1.0
  description: Static OpenAPI draft aligned with docs/specs/api-design.md
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/email/request-code:
    post:
      summary: Request email code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Code requested
  /auth/login:
    post:
      summary: Login with email code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Login success
...
  /search:
    get:
      summary: Unified content search
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: Search results
  /content/{id}:
    get:
      summary: Get content detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content item
  /content/{id}/related:
    get:
      summary: Get related content
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Related items
...
  /library/playlists/{id}/items:
    post:
      summary: Add playlist items
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Items added
...
  /library/favorites/{contentId}:
    delete:
      summary: Remove favorite
      parameters:
        - in: path
          name: contentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Favorite removed
...